+++ lib/gorillib/collection.rb
@@ -2,20 +2,27 @@ require 'gorillib/metaprogramming/delegation'
 require 'gorillib/metaprogramming/class_attribute'
 
 module Gorillib
-  class Collection
-    # [String, Symbol] Method invoked on a new item to generate its collection key; :to_key by default
-    attr_accessor :key_method
-    # The default `key_method` invoked on a new item to generate its collection key
-    DEFAULT_KEY_METHOD = :to_key
 
+  #
+  # A generic collection stores objects uniquely, in the order added. It responds to:
+  #   - receive!, values, to_a, each and each_value;
+  #   - length, size, empty?, blank?
+  #
+  # A Collection additionally lets you store and retrieve things by label:
+  #   - [], []=, include?, fetch, delete, each_pair, to_hash.
+  #
+  # A ModelCollection adds:
+  #   - `key_method`: called on objects to get their key; `to_key` by default.
+  #   - `factory`: generates new objects, converts received objects
+  #   - `<<`: adds object under its `key_method` key
+  #   - `receive!`s an array by auto-keying the elements, or a hash by trusting what you give it
+  #   - `update_or_create: if absent, creates object with given attributes and
+  #     `key_method => key`; if present, updates with given attributes.
+  #
+  class GenericCollection
     # [{Symbol => Object}] The actual store of items, but not for you to mess with
     attr_reader :clxn
     protected   :clxn
-
-    def initialize(key_method=DEFAULT_KEY_METHOD)
-      @clxn        = Hash.new
-      @key_method  = key_method
-    end
 
     def self.receive(items, *args)
       coll = new(*args)
@@ -23,21 +30,7 @@ module Gorillib
       coll
     end
 
-    # common to all collections, delegable to all
-    delegate :[], :[]=, :fetch,                       :to => :clxn
-    delegate :length, :size, :empty?, :blank?,        :to => :clxn
-    # common to all collections, delegable to hash
-    delegate :values, :delete,                        :to => :clxn
-    # novel to a labelled collection
-    delegate :keys, :each_pair, :each_value,          :to => :clxn
-    delegate :include?,                               :to => :clxn
-
-    # Adds an item in-place
-    # @return [Gorillib::Collection] the collection
-    def <<(val)
-      receive! [val]
-      self
-    end
+    delegate :length, :size, :empty?, :blank?, :to => :clxn
 
     # Two collections are equal if they have the same class and their contents are equal
     #
@@ -48,9 +41,36 @@ module Gorillib
       clxn == other.send(:clxn)
     end
 
-    #
-    # Hash-backed collection
-    #
+  public
+
+    # @return [String] string describing the collection's array representation
+    def to_s           ; to_a.to_s           ; end
+    # @return [String] string describing the collection's array representation
+    def inspect(detailed=true)
+      if detailed then  guts = clxn.map{|key, val| "%-15s %s" % ["#{key}:", val.inspect] }.join(",\n   ")
+      else              guts =  keys.join(", ") ; end
+      ["c{ ", guts, " }"].join
+    end
+    # @return [Array] serializable array representation of the collection
+    def to_wire(options={})
+      to_a.map{|el| el.respond_to?(:to_wire) ? el.to_wire(options) : el }
+    end
+    # same as #to_wire
+    def as_json(*args) to_wire(*args) ; end
+    # @return [String] JSON serialization of the collection's array representation
+    def to_json(*args) to_wire(*args).to_json(*args) ; end
+
+  end
+
+  class Collection < Gorillib::GenericCollection
+    def initialize
+      @clxn   = Hash.new
+    end
+
+    delegate :[], :[]=, :fetch, :delete,       :to => :clxn
+    delegate :values,                          :to => :clxn
+    delegate :keys, :each_pair, :each_value,   :to => :clxn
+    delegate :include?,                        :to => :clxn
 
     # @return [Array] an array holding the items
     def to_a    ; values    ; end
@@ -74,38 +94,7 @@ module Gorillib
     # - otherwise, it receives a hash generates from the id/value pairs of each object in the given collection.
     def convert_collection(cc)
       return cc.to_hash if cc.respond_to?(:to_hash)
-      cc.inject({}) do |acc, val|
-        key      = val.public_send(key_method)
-        acc[key] = val
-        acc
-      end
-    end
-
-  public
-
-    # @return [String] string describing the collection's array representation
-    def to_s           ; to_a.to_s           ; end
-    # @return [String] string describing the collection's array representation
-    def inspect(detailed=true)
-      str = "c{ "
-      if detailed
-        str << clxn.map do |key, val|
-          "%-15s %s" % ["#{key}:", val.inspect]
-        end.join(",\n   ")
-      else
-        str << keys.join(", ")
-      end
-      str << " }"
-    end
-    # @return [Array] serializable array representation of the collection
-    def to_wire(options={})
-      to_a.map{|el| el.respond_to?(:to_wire) ? el.to_wire(options) : el }
-    end
-    def as_json(*args) to_wire(*args) ; end
-
-    # @return [String] JSON serialization of the collection's array representation
-    def to_json(*args)
-      to_wire(*args).to_json(*args)
+      raise "a #{self.class} can only receive a hash with explicitly-labelled contents."
     end
 
   end