+++ lib/gorillib/collection/model_collection.rb
@@ -1,24 +1,31 @@
 module Gorillib
   class ModelCollection < Gorillib::Collection
+    # [String, Symbol] Method invoked on a new item to generate its collection key; :to_key by default
+    attr_accessor :key_method
+    # The default `key_method` invoked on a new item to generate its collection key
+    DEFAULT_KEY_METHOD = :to_key
 
     # [Class, #receive] Factory for generating a new collection item.
     class_attribute :factory, :instance_writer => false
     singleton_class.class_eval{ protected :factory= }
 
-    def initialize(factory=Whatever, key_method=DEFAULT_KEY_METHOD)
-      @factory     = Gorillib::Factory(factory)
+    def initialize(key_meth=nil, obj_factory=nil)
+      @factory     = Gorillib::Factory(obj_factory) if obj_factory
       @clxn        = Hash.new
-      @key_method  = key_method
+      @key_method  = key_meth || DEFAULT_KEY_METHOD
+    end
+
+    # Adds an item in-place
+    # @return [Gorillib::Collection] the collection
+    def <<(val)
+      receive! [val]
+      self
     end
 
     def create(*args, &block)
       item = factory.receive(*args)
       self << item
       item
-    end
-
-    def find_or_create(key)
-      fetch(key){ create(key_method => key) }
     end
 
     def update_or_create(key, *args, &block)